chore: combine template and conversation
