+++ src/utils/cmd.py
@@ -1,5 +1,4 @@
-from chatgpt_cli.conversation import Conversation
-from chatgpt_cli.template import Template
+from chatgpt_cli.conversation import Conversation, Template
 from utils.io import *
 
 
@@ -10,9 +9,9 @@ def is_command(user_msg: str) -> bool:
 
 
 def execute_command(
-        user_msg: str,
-        conv: Conversation,
-        tmpl: Template,
+    user_msg: str,
+    conv: Conversation,
+    tmpl: Template,
 ) -> str:
     user_msg = user_msg.strip()
     if user_msg in ["!help", "help"]:
@@ -40,9 +39,7 @@ def execute_command(
         print("Bye!")
         exit(0)
     elif user_msg.startswith("!tmpl"):
-        tmpl.execute_command(conv, user_msg)
-
-
+        tmpl.execute_command(user_msg, conv)
     elif user_msg.startswith("!"):
         print("Invalid command, please try again")
     return user_msg
